import RouteViewer_Enhanced_20100110_src.zip 20100110 (2010-01-12)
This is an enhanced version of the OpenBVE tool Route Viewer. It is
based on the version 1.2.3.3 source code release.

Enhancements:
* Added a button on the messages window to copy the messages to the
  clipboard.
* Added a function to go to any position on the route.
* Changed the key prompts displayed for Page Up and Page Dn to PgUp and
  PgDn.
* Fixed a number of spelling errors.

I first wrote these enhancements over 7 months ago. At that time there
was no way to export the message log from the programme. I added an
option to copy the messages to the clipboard so that I could export them
for further examination. Since then, the author has added a function to
export to a file but having both options is even nicer.

The most tedious thing about viewing routes with the standard Route
Viewer is that you have to drive the train manually to the point of
interest every time. This can be very time consuming when you want to
keep returning to a given point to check changes that you have made to a
route. You can move between stations, which is a big help, and the train
moves faster than it does in the game but the obvious function to move
to a specific point is missing. I have added it. There is a new function
which uses the key "G" and opens a dialogue which allows you to enter a
specific distance and moves the train to that location when OK is
pressed. You can enter the value manually or use the up/down controls
provided. this makes moving around within a route much easier.

The other thing I noticed was that the description for the Page Up and
Page Dn keys was not intuitive. It confused me more than once. I have
therefore changed the displayed descriptions to PgUp and PgDn, which
seems much clearer to me.

Run the executable file from the OpenBVE directory as it needs access to
the data directory.

Enjoy.